+++ release.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+python3 scripts/release.py